generalized setThemeStylesheet
